unit tests for cwhub, split code, minor cleanup
